fix(web): apply full-archive bars-hook fix to the web ladder too (Codex round 2)
Round-2 adversarial review found the web app's ladder bars hook
(_install_web_marketdata_caches) carried the identical window.to_dataframe()
truncation as the pipeline hook. The ticker_detail JIT path uses the web app's
ohlcv_cache.get_or_fetch, so a short-listed ticker could still render sparse
there. Extract the fix into a shared resolve_full_archive_bars helper in the
ladder module and call it from both hooks so the convergence cannot drift
between the two duplicated closures. Add a web-hook convergence regression.

Also normalize compute_chart_source_hash's CRC to big-endian bytes so the
provenance value is identical across platforms of differing native float byte
order.

+++ swing/web/charts.py
@@ -107,7 +107,11 @@ def compute_chart_source_hash(bars: Any) -> str:
     close_crc = 0
     if "Close" in getattr(bars, "columns", []):
         try:
-            close_bytes = bars["Close"].to_numpy(dtype="float64").tobytes()
+            # Big-endian normalization so the CRC is identical across platforms
+            # of differing native float byte order (cross-process determinism).
+            close_bytes = (
+                bars["Close"].to_numpy(dtype="float64").astype(">f8").tobytes()
+            )
             close_crc = zlib.crc32(close_bytes) & 0xFFFFFFFF
         except Exception:  # noqa: BLE001 — provenance is best-effort; degrade to count/window
             close_crc = 0
